refactor(council): extract the parity serialiser and lower the size ratchet
Second and third payback steps for the 408 violations this branch put into three
already-over-ceiling files. The first (cli_call_budget.ts) reached 147; comment
trimming, measured rather than assumed, could not close the rest — of the 408
added lines 241 were comment or blank, so deleting every comment written here
still left real code behind. Both other entries in that baseline note record the
same conclusion, reached independently.

_py_json.ts takes the Python-parity JSON serialiser out of clients.ts verbatim
(147 -> 20). It is a MOVE and must stay one: a frozen byte-parity golden compares
the counter state file against output captured from the deleted Python twin, so
ensure_ascii, insertion-order keys, Python float repr and astral-plane surrogate
pairs are all load-bearing. The file header says so, because the next reader will
otherwise see 85 lines of hand-rolled JSON and reach for JSON.stringify.

quota_summary_line's formatting moves into the budget module (20 -> -21), typed
structurally against the fields it reads rather than importing CliClient, which
keeps the new module free of any dependency on clients.ts.

The baseline goes 19,647 -> 19,626 — a lowering, the only direction this ratchet
allows, with the three steps and their order recorded in its note. 1,018
ai_council + ai_team tests pass untouched, the byte-parity golden included, which
is the evidence that all three moves moved code rather than rewrote it.
